Move index_put regression to Python PT2 tests
