+++ crates/luminal_python/rust/src/translator/mod.rs
@@ -406,84 +406,3 @@ impl<'a> Translator<'a> {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{
-        fs,
-        process::Command,
-        time::{SystemTime, UNIX_EPOCH},
-    };
-
-    use super::*;
-
-    fn input_backed_index_put_program() -> ParsedPT2 {
-        let nonce = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("system time should be after the Unix epoch")
-            .as_nanos();
-        let path = std::env::temp_dir().join(format!(
-            "luminal_input_backed_index_put_{}_{}.pt2",
-            std::process::id(),
-            nonce
-        ));
-        let script = r#"
-import sys
-import torch
-
-
-class InputBackedIndexPut(torch.nn.Module):
-    def forward(self, cache, positions, values):
-        cache[:, positions] = values
-        return cache
-
-
-cache = torch.zeros(2, 4)
-positions = torch.tensor([1], dtype=torch.int64)
-values = torch.ones(2, 1)
-exported = torch.export.export(
-    InputBackedIndexPut(),
-    (cache, positions, values),
-    strict=False,
-)
-torch.export.save(exported, sys.argv[1])
-"#;
-
-        let python = std::env::var_os("PYTHON").unwrap_or_else(|| "python3".into());
-        let output = Command::new(python)
-            .arg("-c")
-            .arg(script)
-            .arg(&path)
-            .output()
-            .expect("Python should be available to export the PT2 fixture");
-        assert!(
-            output.status.success(),
-            "PyTorch export failed:\n{}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-
-        let parsed = crate::pt2_parser::parse_pt2(
-            path.to_str()
-                .expect("temporary PT2 path should contain valid UTF-8"),
-        )
-        .expect("PyTorch-exported PT2 program should parse");
-        fs::remove_file(path).expect("temporary PT2 fixture should be removable");
-        parsed
-    }
-
-    #[test]
-    fn input_backed_index_put_output_aliases_destination_input() {
-        let translated = translate(&input_backed_index_put_program())
-            .expect("input-backed index_put should translate");
-
-        let (input_name, input_id) = &translated.user_input_ids[0];
-        let (output_name, output_id) = &translated.output_ids[0];
-
-        assert_eq!(input_name, "cache");
-        assert!(output_name.contains("index_put"));
-        assert_eq!(
-            output_id, input_id,
-            "an index_put write-back into a user input must expose that input as its output"
-        );
-    }
-}
